MAYA-106529 - MayaUSD: Cleanup old UFE/MayaPR conditionals
* Put "test" back in file "testDeleteCmd.py" and removed new
  "testDeleteCmd_v2.py". Created new ufeUtils.ufeFeatureSetVersion()
  for use by tests when a v2 (or greater) check is needed.
* Removed old MayaPR checks.
