+++ test/lib/ufe/testDeleteCmd.py
@@ -18,7 +18,7 @@
 
 import maya.cmds as cmds
 
-from ufeTestUtils import usdUtils, mayaUtils
+from ufeTestUtils import ufeUtils, usdUtils, mayaUtils
 import ufe
 
 import unittest
@@ -54,7 +54,7 @@ class DeleteCmdTestCase(unittest.TestCase):
     UFE Feature : SceneItemOps
     Maya Feature : delete
     Action : Remove object from scene.
-    Applied On Selection :
+    Applied On Selection / Command Arguments:
         - Multiple Selection [Mixed, Non-Maya].  Maya-only selection tested by
           Maya.
     Undo/Redo Test : Yes
@@ -185,3 +185,115 @@ class DeleteCmdTestCase(unittest.TestCase):
         # After the undo we again added two items back.
         self.assertEqual(ufeObs.nbDeleteNotif(), 4)
         self.assertEqual(ufeObs.nbAddNotif(), 4)
+
+    @unittest.skipUnless(ufeUtils.ufeFeatureSetVersion() >= 2, 'testDeleteArgs only available in UFE v2 or greater.')
+    def testDeleteArgs(self):
+        '''Delete Maya and USD objects passed as command arguments.'''
+
+        # Create our UFE notification observer
+        ufeObs = TestObserver()
+        ufe.Scene.addObjectDeleteObserver(ufeObs)
+        ufe.Scene.addObjectAddObserver(ufeObs)
+
+        spherePath = ufe.Path(mayaUtils.createUfePathSegment("|pSphere1"))
+        sphereItem = ufe.Hierarchy.createItem(spherePath)
+        sphereShapePath = ufe.Path(
+            mayaUtils.createUfePathSegment("|pSphere1|pSphereShape1"))
+        sphereShapeItem = ufe.Hierarchy.createItem(sphereShapePath)
+
+        mayaSegment = mayaUtils.createUfePathSegment(
+            "|world|transform1|proxyShape1")
+        ball35Path = ufe.Path(
+            [mayaSegment,
+             usdUtils.createUfePathSegment("/Room_set/Props/Ball_35")])
+        ball35Item = ufe.Hierarchy.createItem(ball35Path)
+        ball34Path = ufe.Path(
+            [mayaSegment,
+             usdUtils.createUfePathSegment("/Room_set/Props/Ball_34")])
+        ball34Item = ufe.Hierarchy.createItem(ball34Path)
+        propsPath = ufe.Path(
+            [mayaSegment, usdUtils.createUfePathSegment("/Room_set/Props")])
+        propsItem = ufe.Hierarchy.createItem(propsPath)
+
+        sphereShapeName = str(sphereShapeItem.path().back())
+        ball35Name = str(ball35Item.path().back())
+        ball34Name = str(ball34Item.path().back())
+
+        # Before delete, each item is a child of its parent.
+        sphereHierarchy = ufe.Hierarchy.hierarchy(sphereItem)
+        propsHierarchy = ufe.Hierarchy.hierarchy(propsItem)
+
+        sphereChildren = sphereHierarchy.children()
+        propsChildren = propsHierarchy.children()
+
+        sphereChildrenNames = childrenNames(sphereChildren)
+        propsChildrenNames = childrenNames(propsChildren)
+
+        self.assertIn(sphereShapeItem, sphereChildren)
+        self.assertIn(ball35Item, propsChildren)
+        self.assertIn(sphereShapeName, sphereChildrenNames)
+        self.assertIn(ball35Name, propsChildrenNames)
+
+        ball34PathString = ufe.PathString.string(ball34Path)
+        self.assertEqual(
+            ball34PathString,
+            "|transform1|proxyShape1,/Room_set/Props/Ball_34")
+
+        # Test that "|world" prefix is optional for multi-segment paths.
+        ball35PathString = "|transform1|proxyShape1,/Room_set/Props/Ball_35"
+
+        ufeObs.reset()
+        cmds.delete(
+            ball35PathString, ball34PathString, "|pSphere1|pSphereShape1")
+
+        # We deleted 3 items.
+        self.assertEqual(ufeObs.nbDeleteNotif(), 3)
+        self.assertEqual(ufeObs.nbAddNotif(), 0)
+
+        sphereChildren = sphereHierarchy.children()
+        propsChildren = propsHierarchy.children()
+
+        sphereChildrenNames = childrenNames(sphereChildren)
+        propsChildrenNames = childrenNames(propsChildren)
+
+        self.assertNotIn(sphereShapeName, sphereChildrenNames)
+        self.assertNotIn(ball35Name, propsChildrenNames)
+        self.assertNotIn(ball34Name, propsChildrenNames)
+        self.assertFalse(cmds.objExists("|pSphere1|pSphereShape1"))
+
+        cmds.undo()
+
+        # After the undo we added three items back.
+        self.assertEqual(ufeObs.nbDeleteNotif(), 3)
+        self.assertEqual(ufeObs.nbAddNotif(), 3)
+
+        sphereChildren = sphereHierarchy.children()
+        propsChildren = propsHierarchy.children()
+
+        sphereChildrenNames = childrenNames(sphereChildren)
+        propsChildrenNames = childrenNames(propsChildren)
+
+        self.assertIn(sphereShapeItem, sphereChildren)
+        self.assertIn(ball35Item, propsChildren)
+        self.assertIn(ball34Item, propsChildren)
+        self.assertIn(sphereShapeName, sphereChildrenNames)
+        self.assertIn(ball35Name, propsChildrenNames)
+        self.assertIn(ball34Name, propsChildrenNames)
+        self.assertTrue(cmds.objExists("|pSphere1|pSphereShape1"))
+
+        cmds.redo()
+
+        # After the redo we again deleted three items.
+        self.assertEqual(ufeObs.nbDeleteNotif(), 6)
+        self.assertEqual(ufeObs.nbAddNotif(), 3)
+
+        sphereChildren = sphereHierarchy.children()
+        propsChildren = propsHierarchy.children()
+
+        sphereChildrenNames = childrenNames(sphereChildren)
+        propsChildrenNames = childrenNames(propsChildren)
+
+        self.assertNotIn(sphereShapeName, sphereChildrenNames)
+        self.assertNotIn(ball35Name, propsChildrenNames)
+        self.assertNotIn(ball34Name, propsChildrenNames)
+        self.assertFalse(cmds.objExists("|pSphere1|pSphereShape1"))
