+++ test/lib/ufe/testMayaPickwalk.py
@@ -168,7 +168,7 @@ class MayaUFEPickWalkTesting(unittest.TestCase):
         
         # Pickwalk on unsupported UFE items
         cmds.select("pCube1.e[6]")
-        if int(cmds.about(majorVersion=True)) <= 2020 or mayaUtils.previewReleaseVersion() < 115:
+        if int(cmds.about(majorVersion=True)) <= 2020:
             self.snapShotAndTest((["pCube1.e[6]"], ["pCubeShape1"]))
             self.assertTrue(next(iter(ufe.GlobalSelection.get())).isProperty())
         else:
