+++ test/lib/ufe/testScaleCmd.py
@@ -171,7 +171,6 @@ class ScaleCmdTestCase(testTRSBase.TRSTestCaseBase):
 
         self.runTestScale(expected)
 
-    @unittest.skipUnless(mayaUtils.previewReleaseVersion() >= 115, 'Requires Maya fixes only available in Maya Preview Release 115 or later.') 
     def testScaleUSD(self):
         '''Scale USD object, read through the Transform3d interface.'''
 
@@ -208,7 +207,6 @@ class ScaleCmdTestCase(testTRSBase.TRSTestCaseBase):
 
         self.runTestScale(expected)
 
-    @unittest.skipUnless(mayaUtils.previewReleaseVersion() >= 115, 'Requires Maya fixes only available in Maya Preview Release 115 or later.') 
     def testMultiSelectScaleUSD(self):
         '''Scale multiple USD objects, read through Transform3d interface.'''
