+++ test/lib/ufe/testSelection.py
@@ -127,7 +127,6 @@ class SelectTestCase(unittest.TestCase):
             ufeSelectCmd.replaceWith(sn)
         self.runTestSelection(selectCmd)
 
-    @unittest.skipUnless(mayaUtils.previewReleaseVersion() >= 116, 'Requires Maya fixes only available in Maya Preview Release 116 or later.') 
     def testMayaSelect(self):
         # At time of writing (17-May-2020), Maya select command does not
         # accept UFE path strings.  Use Maya select for Maya scene items,
